pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["aristotle_mcp*"]

[project]
name = "aristotle-mcp"
version = "0.1.0"
description = "Git-based rule management MCP server for Aristotle"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.27.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0",
]

[project.scripts]
aristotle-mcp = "aristotle_mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["test"]
